Basic Database Functionality
Created the ability to register new users, and login. Added a basic User class. Created a database. This is not hooked into the site but the functions can be called, and they can add to the database and check if the correct password was entered. There is a bit of example code at the bottom showing how this can be interfaced with.